修复发工资前核查发现的一批严重问题：午间打卡窗口重复导致整天工时清零、审批事务与重试策略冲突导致审批全部失败、加班重复提交导致工时多算、请假天数边界情况与半天假出勤重复计算
- 班次午间必打卡窗口去重（ShiftSchedule/AttendanceRecord 解析层 + ShiftManage 保存校验），修正 GA 员工白班/夜班配置重复窗口导致整天工时被清零的问题
- ApprovalService.HandleApprovalAsync 补上 CreateExecutionStrategy 包裹事务，修复审批通过/驳回全部报错失败的问题
- 提交申请时新增同类型时间段重叠检查，阻止重复提交；ApplySubmit 页提交按钮点击后立即锁住
- 远程打卡每日次数上限从 20/40 收紧为 6/12（阿里云人脸识别按次计费），并修复次数用满导致下班打卡被挡的问题
- 请假区间与当天无实际交集（如结束时间恰好卡在午夜）时跳过，不再产生状态为请假但时长为 0 的记录
- 月度汇总"实际出勤天数"改为按请假占比折算，半天假不再跟出勤天数重复计满整天；ActualWorkdays 由 int 改为 decimal（含 EF 迁移）

+++ Helpers/ExcelExportHelper.cs
@@ -67,7 +67,7 @@ public static class ExcelExportHelper
             SetCell(row, 2,  dto.DeptName ?? "",                          baseStyle);
             SetCell(row, 3,  dto.Position ?? "",                          baseStyle);
             SetCell(row, 4,  dto.ExpectedWorkdays,                        baseStyle);
-            SetCell(row, 5,  dto.ActualWorkdays,                          baseStyle);
+            SetCell(row, 5,  (double)dto.ActualWorkdays,                  baseStyle);
             SetCell(row, 6,  dto.NightShiftDays,                          baseStyle);   // 夜班天数
             SetCell(row, 7,  dto.LateCount,        dto.LateCount > 0        ? orangeStyle : baseStyle);  // 有迟到→橙
             SetCell(row, 8,  dto.EarlyLeaveCount,  dto.EarlyLeaveCount > 0  ? orangeStyle : baseStyle);  // 有早退→橙
@@ -86,7 +86,7 @@ public static class ExcelExportHelper
             totalRow.HeightInPoints = 18;
             SetCell(totalRow, 0,  "合计",                                                     headerStyle);
             SetCell(totalRow, 4,  summaries.Sum(s => s.ExpectedWorkdays),                    headerStyle);
-            SetCell(totalRow, 5,  summaries.Sum(s => s.ActualWorkdays),                      headerStyle);
+            SetCell(totalRow, 5,  (double)summaries.Sum(s => s.ActualWorkdays),              headerStyle);
             SetCell(totalRow, 6,  summaries.Sum(s => s.NightShiftDays),                      headerStyle);
             SetCell(totalRow, 7,  summaries.Sum(s => s.LateCount),                           headerStyle);
             SetCell(totalRow, 8,  summaries.Sum(s => s.EarlyLeaveCount),                     headerStyle);
